// src/app/model/raw.model.ts
export class RawMaterial{
  id: number | undefined;
  name: string | undefined;
  code: string | undefined;
  description: string | undefined;
  quantity: number | undefined;
  unit: string | undefined;
  pricePerUnit: number | undefined;
  expiryDate: Date | undefined;
  batchNumber: string | undefined;
  supplierName: string | undefined;
  supplierId: number | undefined;
  purchaseDate: Date | undefined;  
  storageLocation: string | undefined;
  status: string | undefined;
   
}