// src/app/model/rawimport.model.ts
export class RawMaterialImport {
  id: number | undefined;
  sellerId: number | undefined;
  materialId: number | undefined;
  sellerName: string | undefined;
  materialName: string | undefined;
  quantity: number | undefined;
  unitPrice: number | undefined;
  totalPrice: number | undefined;
  importDate: Date | undefined;
  
}